Update after run training and test methods
