--- conv2d/app/requirements.txt
@@ -1,7 +1,0 @@
-lxml
-h5py
-numpy
-pyshark
-scikit-learn
-tensorflow
-jupyter
