--- conv2d/app/sample-dataset/10t-10n-SYN2020-LUCID.csv
@@ -1,2 +1,0 @@
-Model         TIME(sec) ACC    ERR    F1     PPV    TPR    FPR    TNR    FNR    Parameters
-SYN2020-LUCID     0.212 1.0000 0.6148 1.0000 1.0000 1.0000 0.0000 1.0000 0.0000 lr=0.100,b=1024,n=010,t=010,k=002,h=(03,11),m=max
